+++ CHANGELOG.md
@@ -2896,3 +2896,4 @@
 - [2026-07-16 14:15] style(test): harmonize light and dark theme palette for test
 - [2026-07-16 14:53] chore(theme): update dependencies and cleanup unused imports in theme
 - [2026-07-16 14:55] docs(state): update readme and API contracts for state
+- [2026-07-16 15:29] fix(state): prevent double tap duplicate action in state